pyproject.toml
[tool.poetry]
name = "bctqr"
version = "0.1.0"
description = ""
authors = ["Frederic Vogels <frederic.vogels+github@gmail.com>"]
readme = "README.md"

[tool.poetry.group.dev.dependencies]
mypy = "^1.5.1"

[tool.setuptools]
include_package_data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
mypkg = ["*.png"]

[tool.poetry.dependencies]
python = "^3.11"
reportlab = {extras = ["pycairo"], version = "^4.0.4"}
click = "^8.1.7"
qrcode = "^7.4.2"
pillow = "^10.0.0"
pylabels = "^1.2.1"
pydantic = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
bctqr = 'bctqr.cli:cli'
